update sdk test for encrypt sql && rewrite buffer
