+++ src/main/java/dao/SQLBuffer.java
@@ -1,0 +1,12 @@
+package dao;
+
+import io.vertx.core.buffer.Buffer;
+
+public class SQLBuffer {
+    public Buffer buffer;
+    public String queryId;
+    public SQLBuffer(Buffer buffer, String queryId){
+        this.buffer=buffer;
+        this.queryId=queryId;
+    }
+}
