+++ src/main/java/dao/SQLResponse.java
@@ -1,0 +1,7 @@
+package dao;
+
+public class SQLResponse {
+    public String encryptSql;
+    public String table;
+    public String queryId;
+}
